remove protocol.py
