+++ test/ut/sdk/test_assessor.py
@@ -8,10 +8,7 @@ from unittest import TestCase, main
 from nni.assessor import Assessor, AssessResult
 from nni.runtime import msg_dispatcher_base as msg_dispatcher_base
 from nni.runtime.msg_dispatcher import MsgDispatcher
-from nni.runtime import protocol
-from nni.runtime.protocol import CommandType, send, receive
-
-from .helper.legacy_command_channel import LegacyCommandChannel
+from nni.runtime.tuner_command_channel.legacy import *
 
 _trials = []
 _end_trials = []
@@ -36,15 +33,15 @@ _out_buf = BytesIO()
 def _reverse_io():
     _in_buf.seek(0)
     _out_buf.seek(0)
-    protocol._set_out_file(_in_buf)
-    protocol._set_in_file(_out_buf)
+    _set_out_file(_in_buf)
+    _set_in_file(_out_buf)
 
 
 def _restore_io():
     _in_buf.seek(0)
     _out_buf.seek(0)
-    protocol._set_in_file(_in_buf)
-    protocol._set_out_file(_out_buf)
+    _set_in_file(_in_buf)
+    _set_out_file(_out_buf)
 
 
 class AssessorTestCase(TestCase):
